+++ tests/unit/test_auth.py
@@ -3,7 +3,7 @@ from http import HTTPStatus
 
 from mock import patch
 from biostudiesclient.auth import Auth
-from biostudiesclient.rest_error_exception import RestErrorException
+from biostudiesclient.exceptions import RestErrorException
 
 
 class TestAuth(unittest.TestCase):
